+++ src/types/MakeInfo.ts
@@ -1,8 +1,18 @@
+import { standardOpenOCDInterface } from '../Definitions';
+
+export interface Stm32SettingsInterface {
+  armToolchainPath: string;
+  openOCDPath: string;
+  makePath: string;
+  openOCDInterface: string;
+}
+
 export interface ToolChainInterface {
-  openOCD: string | boolean;
-  make: string | boolean;
-  cMake: string | boolean;
-  armToolchain: string | boolean;
+  openOCDPath: string | boolean;
+  makePath: string | boolean;
+  cMakePath: string | boolean;
+  armToolchainPath: string | boolean;
+  openOCDInterface: string;
 }
 
 export interface BuildFilesInterface {
@@ -13,6 +23,8 @@ export interface BuildFilesInterface {
   libs: string[];
   libDirs: string[];
 }
+
+export type STM32Languages = 'C' | 'C++';
 export interface MakeInfoInterface extends BuildFilesInterface {
   cDefs: string[];
   cxxDefs: string[];
@@ -25,13 +37,15 @@ export interface MakeInfoInterface extends BuildFilesInterface {
   mcu: string;
   ldscript: string;
   targetMCU: string;
+  language: STM32Languages;
 }
 
 export class ToolChain implements ToolChainInterface {
-  public openOCD: string | boolean = false;
-  public make: string | boolean = false;
-  public cMake: string | boolean = false;
-  public armToolchain: string | boolean = false;
+  public openOCDPath: string | boolean = false;
+  public makePath: string | boolean = false;
+  public cMakePath: string | boolean = false;
+  public armToolchainPath: string | boolean = false;
+  public openOCDInterface = standardOpenOCDInterface;
   // public constructor() { }
 }
 export class BuildFiles implements BuildFilesInterface {
@@ -62,5 +76,6 @@ export default class MakeInfo implements MakeInfoInterface {
   public mcu = '';
   public ldscript = '';
   public targetMCU = '';
+  public language = 'C' as STM32Languages;
   // public constructor() { }
 }