+++ src/workspaceConfiguration/LaunchTasksConfig.ts
@@ -1,7 +1,6 @@
-import getOpenOCDTarget from '../OpenOcdTargetFiles';
 import MakefileInfo from '../types/MakeInfo';
 import { TaskDefinition } from 'vscode';
-
+import getOpenOCDTarget from '../OpenOcdTargetFiles';
 
 export default function getLaunchTask(info: MakefileInfo): TaskDefinition {
   const config = {
@@ -16,7 +15,7 @@ export default function getLaunchTask(info: MakefileInfo): TaskDefinition {
     preLaunchTask: 'Build STM',
     device: 'stlink',
     configFiles: [
-      'interface/stlink.cfg',
+      info.tools.openOCDInterface,
       `target/${getOpenOCDTarget(info.targetMCU)}`,
     ],
   };